+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix the SMTP receiver being unable to receive mail from MailGun.
 - Fix panic when attempting to configure an invalid key pattern if Crymap was
   built with more recent Rust versions.
+- Fix incorrect syntax of Return-Path headers added by `smtpin` service.
 
 # 2.0.0